+++ database/migrations/2025_11_03_120631_create_production_plan_step3_table.php
@@ -18,14 +18,14 @@ return new class extends Migration
             $table->foreignId('production_plan_id')->constrained('production_plans')->onDelete('cascade');
             $table->foreignId('gelondongan_item_id')->constrained('items')->onDelete('restrict');
             $table->foreignId('kerupuk_kering_item_id')->constrained('items')->onDelete('restrict');
-            $table->decimal('qty_gl1_gelondongan', 10, 3)->default(0);
-            $table->decimal('qty_gl1_kg', 10, 3)->default(0);
-            $table->decimal('qty_gl2_gelondongan', 10, 3)->default(0);
-            $table->decimal('qty_gl2_kg', 10, 3)->default(0);
-            $table->decimal('qty_ta_gelondongan', 10, 3)->default(0);
-            $table->decimal('qty_ta_kg', 10, 3)->default(0);
-            $table->decimal('qty_bl_gelondongan', 10, 3)->default(0);
-            $table->decimal('qty_bl_kg', 10, 3)->default(0);
+            $table->integer('qty_gl1_gelondongan')->default(0);
+            $table->decimal('qty_gl1_kg', 10, 2)->default(0);
+            $table->integer('qty_gl2_gelondongan')->default(0);
+            $table->decimal('qty_gl2_kg', 10, 2)->default(0);
+            $table->integer('qty_ta_gelondongan')->default(0);
+            $table->decimal('qty_ta_kg', 10, 2)->default(0);
+            $table->integer('qty_bl_gelondongan')->default(0);
+            $table->decimal('qty_bl_kg', 10, 2)->default(0);
             $table->timestamps();
 
             $table->index('production_plan_id');
